Migrated WorldCities to unified api
And updated some namespace references

+++ WorldCities/WorldCities/WorldCitiesListController.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 
 namespace WorldCities
 {
@@ -55,7 +55,7 @@ namespace WorldCities
 				this.controller = controller;
 			}
 			
-			public override int RowsInSection (UITableView tableview, int section)
+			public override nint RowsInSection (UITableView tableview, nint section)
 			{
 				return controller.cityList.Count; 
 			}
